Updating (#12)
* Added notes about the Google Assistant integration being broken

* This fixes the issue of a "stale" CGM value displaying indefinately. It displays "Internet offline." until internet is restored.

This fix addresses the issue that with a lost internet connection, clock view continues to display a stale CGM values forever, with no clue that internet connectivity is lost. This fix shows "Internet offline." for the duration that is it lost, then returns to normal when internet is restored.

* create room for dev post release

* New translations en.json (Italian)

* New translations en.json (Italian)

* Update es_ES.json

Add "virtAsstLaunch"

* New translations en.json (Turkish)

* New translations en.json (Portuguese)

* New translations en.json (Russian)

* New translations en.json (Arabic)

* New translations en.json (Chinese Traditional)

* New translations en.json (Chinese Traditional)

* New translations en.json (Chinese Traditional)

* New translations en.json (Chinese Traditional)

* New translations en.json (Chinese Traditional)

* New translations en.json (Chinese Traditional)

* New translations en.json (Chinese Traditional)

* New translations en.json (Chinese Traditional)

* New translations en.json (Chinese Traditional)

* New translations en.json (Chinese Traditional)

* New translations en.json (Chinese Traditional)

* New translations en.json (Chinese Traditional)

* New translations en.json (Chinese Traditional)

* New translations en.json (Chinese Traditional)

* New translations en.json (Chinese Traditional)

* New translations en.json (Chinese Traditional)

* New translations en.json (Chinese Traditional)

* New translations en.json (Chinese Traditional)

* return generic '500' status on bootError page

* Update sensorage.js for 10 day sensors

Changing default SAGE warn and urgent threshold defaults to be useful for 10 day sensors rather than 7 day sensors from the G5 era. SAGE pill turns yellow at day 9 and red 4 hours before expiration of 10 days.

Once G6 is retired, consider setting WARN to 10 days and URGENT to 10.5 days.

* New translations en.json (Spanish)

* Update loop.js

Updated to stop Nightscout crashing if APNs response is not received well from Apple.

* nice error handling, @AndyLow91

* Revert "Update sensorage.js for 10 day sensors"

This reverts commit cec3eebc406bf1f39a39256ac2b1efd850d4e633.
Need to update tests?  Not immediately clear why tests updated from 6 days to 9
days fails, so revertinig for now.

* Fix missing mmol unit conversion for lastEnacted.bg

* allow tests to pass

Switching the way units is fetched to this way makes it similar to the other
code already in place that does something similar.  This allows the tests to
pass with the existing fixtures.
Considering augmenting  additional tests that explicitly tests whether the
relevant branch of code is taken.

* permit any LTS version of node

* continuously test on lts also

This adds the latest long term support version of node to the automated testing
matrix.

* setup-node action syntax is lts/*

---------

Co-authored-by: Caleb <inventor96@users.noreply.github.com>
Co-authored-by: inventor96 <inventor96@gmail.com>
Co-authored-by: gth001 <gth001@gmail.com>
Co-authored-by: Ben West <bewest@gmail.com>
Co-authored-by: Sulka Haro <sulka@sulka.net>
Co-authored-by: antoniomuniz <munizg.antonio@gmail.com>
Co-authored-by: Daniel Cosby <daniel@northflank.com>
Co-authored-by: Ken Fox <46848705+Nightfoxy@users.noreply.github.com>
Co-authored-by: Andy Low <42089621+AndyLow91@users.noreply.github.com>
Co-authored-by: dsnallfot <72826201+dsnallfot@users.noreply.github.com>